refactor submission form and css
