+++ bikespace_frontend/gatsby-browser.ts
@@ -1,0 +1,1 @@
+import './src/styles/global.scss';
