+++ bikespace_frontend/src/components/submission/index.ts
@@ -1,8 +1,16 @@
+export {SubmissionHeader} from './submission-header';
+
+export {SubmissionProgressBar} from './submission-progress-bar';
+
 export {Issue} from './issue';
 export {Location} from './location';
 export {Time} from './time';
 export {Comments} from './comments';
 export {Summary} from './summary';
 export {FeedbackMailTo} from './feedback-mail-to';
-export {SubmissionProgressBar} from './submission-progress';
-export {SubmissionHeader} from './submission-header';
+
+export {SubmissionFormContent} from './submission-form-content';
+export {SubmissionFormController} from './submission-form-controller';
+export {SubmissionForm} from './submission-form';
+
+export {SubmissionPage} from './submission-page';
