+++ bikespace_frontend/src/components/submission/submission-form-content/index.ts
@@ -1,0 +1,1 @@
+export {SubmissionFormContent} from './SubmissionFormContent';
