+++ bikespace_frontend/src/components/submission/submission-form-controller/index.ts
@@ -1,0 +1,1 @@
+export {SubmissionFormController} from './SubmissionFormController';
