+++ bikespace_frontend/src/components/submission/submission-form/index.ts
@@ -1,0 +1,1 @@
+export {SubmissionForm} from './SubmissionForm';
