+++ bikespace_frontend/src/components/submission/submission-page/SubmissionPage.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {SubmissionHeader} from '../submission-header';
+import {SubmissionForm} from '../submission-form';
+
+import * as styles from './submission-page.module.scss';
+
+export function SubmissionPage() {
+  return (
+    <div className={styles.submissionPage}>
+      <SubmissionHeader />
+      <main>
+        <SubmissionForm />
+      </main>
+    </div>
+  );
+}
