+++ bikespace_frontend/src/components/submission/submission-page/index.ts
@@ -1,0 +1,1 @@
+export {SubmissionPage} from './SubmissionPage';
