+++ bikespace_frontend/src/components/submission/submission-page/submission-page.module.scss
@@ -1,0 +1,11 @@
+.submissionPage {
+  position: relative;
+
+  main {
+    position: relative;
+    padding: 20px;
+    background-image: url('@/images/bikespace-background-desktop.jpeg');
+    background-size: cover;
+    background-position: center;
+  }
+}
