+++ bikespace_frontend/src/components/submission/submission-progress-bar/index.ts
