+++ bikespace_frontend/src/components/submission/submission-progress-bar/submission-progress-bar.module.scss
